Moved send success page to a separate file
